Code refactoring

+++ ConnectingDBToApp/Models/TestResult.cs
@@ -2,13 +2,13 @@
 
 public partial class TestResult
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
 
     public string Username { get; set; } = null!;
 
-    public long CountRightAnswer { get; set; }
+    public int CountRightAnswer { get; set; }
 
-    public long CountQuestions { get; set; }
+    public int CountQuestions { get; set; }
 
     public double Percentages { get; set; }
 }
